+++ routing/index.js
@@ -1,3 +1,4 @@
 const routeFromChannels = require('./route_from_channels');
+const routeFromHops = require('./route_from_hops');
 
-module.exports = {routeFromChannels};
+module.exports = {routeFromChannels, routeFromHops};