+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Versions
 
+## 1.6.0
+
+- `routeFromChannels`: Add support for `messages` in routes
+- `routeFromHops`: Add method to calculate a route from a set of hops
+
 ## 1.5.2
 
 - Add `routeFromChannels` method to calculate a route from channels